backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
paho-mqtt>=1.6.1
channels>=4.0
channels-redis>=4.2
daphne>=4.0
django-cors-headers>=4.4.0